refactor: About as global component
