+++ src/components/About/index.tsx
@@ -1,4 +1,4 @@
-import { Button } from '../../../../components/Button'
+import { Button } from '../Button'
 import armchair from '/assets/images/armchair.png'
 
 import * as S from './styles'
