+++ src/components/About/styles.ts
@@ -2,17 +2,16 @@ import styled from 'styled-components'
 
 export const Container = styled.section`
     display: flex;
-    margin: 0 auto;
+    margin: 0px auto;
     gap: 12px;
     max-width: 1440px;
-    width: 90%;
-    padding: 3.5rem 0;
+    padding: 2.8rem 3.8rem;
 
     .card-about {
         display: flex;
         flex-direction: column;
         gap: 12px;
-        padding: 4rem; /* No figma está 3rem(48px) */
+        padding: 4rem; /* No figma está 3rem (48px) */
         background-color: ${(props) => props.theme['dark-primary']};
         color: ${(props) => props.theme.white};
         width: 50%;
@@ -64,7 +63,6 @@ export const Container = styled.section`
     }
 
     @media screen and (max-width: 565px) {
-        padding: 1rem;
 
         .card-about {
             width: 95%;
