+++ src/pages/Home/index.tsx
@@ -4,7 +4,7 @@ import { Header } from '../../components/Header'
 import { Hero } from './components/Hero'
 import { Features } from '../../components/Features'
 import { Products } from '../../components/Products'
-import { About } from './components/About'
+import { About } from '../../components/About'
 import { Signup } from './components/Signup'
 import { Footer } from './components/Footer'
